✅ SeleccionarProducto_Cuando_Dinero_Ingresado_Es_Mayor_Al_Precio_Y_Hay_Cambio_Retorna_Venta_Con_Vueltas - hice el commit sin verificar la prueba pasando :)

+++ TddKatas.VendingMachine/Coin.cs
@@ -9,8 +9,8 @@ public enum Coin
 
 public static class CoinExtensions
 {
-    public static double Valor(this Coin moneda)
+    public static decimal Valor(this Coin moneda)
     {
-        return (int)moneda / 100.0;
+        return (int)moneda / 100m;
     }
 }
